2026-08-10 Preserve float64 precision for long-baseline absolute times
sajax/planet.py hard-cast times to float32 before computing planet positions, which leads to degraded precision at BJD-scale epochs (~2.46e6) and corrupts the time - t_peri cancellation in the mean anomaly.
Times now follow the ambient jax_enable_x64 setting instead of being force-cast, build_transit_model warns when float32 rounding is a significant fraction of the sampling cadence and x64 isn't enabled, and docs/usage.md documents how to opt into float64 for long-baseline datasets.

+++ sajax/core.py
@@ -1032,9 +1032,9 @@ def build_system(
             offsets += dt / (2.0 * oversample)                          # centre sub-bins
             times_oversampled = (
                 times_arr_full[:, None] + offsets[None, :]
-            ).ravel().astype(np.float32)
+            ).ravel()
         else:
-            times_oversampled = times_arr_full.astype(np.float32)
+            times_oversampled = times_arr_full
 
         # ---- Validate k against the wavelength grid (scalar or per-wavelength)
         k_arr = np.atleast_1d(np.asarray(k, dtype=np.float32))
